Consistent Layout & Code Cleanup
Worked on a better, more complete U.I. Some features aren't all done yet and the new layout is only for index.html but the changes will be carried over in the next update. Cleaned up a lot of code and planning to make it more consistent. Fixed U.I Issues in the footer, new settings button (still being worked on) and some other changes.

The layout will be deployed to all pages in the next update with hopefully the new, complete, settings button.